Move ME constants to `constants/mystery-encounters.ts`
