+++ src/data/mystery-encounters/encounters/an-offer-you-cant-refuse-encounter.ts
@@ -23,7 +23,7 @@ import { speciesStarterCosts } from "#app/data/starters";
 import { MysteryEncounterTier } from "#enums/mystery-encounter-tier";
 import { MysteryEncounterOptionMode } from "#enums/mystery-encounter-option-mode";
 import { ModifierRewardPhase } from "#app/phases/modifier-reward-phase";
-import { CLASSIC_MODE_MYSTERY_ENCOUNTER_WAVES } from "#app/constants/constants";
+import { CLASSIC_MODE_MYSTERY_ENCOUNTER_WAVES } from "#app/constants/mystery-encounters";
 import i18next from "i18next";
 
 /** the i18n namespace for this encounter */
