+++ src/data/mystery-encounters/encounters/bug-type-superfan-encounter.ts
@@ -49,7 +49,7 @@ import i18next from "i18next";
 import { MoveInfoOverlay } from "#app/ui/components/move-info-overlay";
 import { allMoves } from "#app/data/data-lists";
 import { ModifierTier } from "#enums/modifier-tier";
-import { CLASSIC_MODE_MYSTERY_ENCOUNTER_WAVES } from "#app/constants/constants";
+import { CLASSIC_MODE_MYSTERY_ENCOUNTER_WAVES } from "#app/constants/mystery-encounters";
 import { getSpriteKeysFromSpecies } from "#app/data/mystery-encounters/utils/encounter-pokemon-utils";
 import { allTrainerConfigs } from "#app/data/trainer-configs/all-trainer-configs";
 import { GAME_WIDTH } from "#app/constants/ui";
