+++ src/data/mystery-encounters/encounters/dark-deal-encounter.ts
@@ -19,7 +19,7 @@ import { MysteryEncounterTier } from "#enums/mystery-encounter-tier";
 import { MysteryEncounterOptionMode } from "#enums/mystery-encounter-option-mode";
 import { ModifierRewardPhase } from "#app/phases/modifier-reward-phase";
 import type { PokemonHeldItemModifier } from "#app/modifier/modifier";
-import { CLASSIC_MODE_MYSTERY_ENCOUNTER_WAVES } from "#app/constants/constants";
+import { CLASSIC_MODE_MYSTERY_ENCOUNTER_WAVES } from "#app/constants/mystery-encounters";
 import { Challenges } from "#enums/challenges";
 import { SpeciesGroups } from "#enums/pokemon-species-groups";
