+++ src/data/mystery-encounters/encounters/shady-vitamin-dealer-encounter.ts
@@ -26,7 +26,7 @@ import { MysteryEncounterTier } from "#enums/mystery-encounter-tier";
 import { MysteryEncounterOptionMode } from "#enums/mystery-encounter-option-mode";
 import type { Nature } from "#enums/nature";
 import { getNatureName } from "#app/data/nature";
-import { CLASSIC_MODE_MYSTERY_ENCOUNTER_WAVES } from "#app/constants/constants";
+import { CLASSIC_MODE_MYSTERY_ENCOUNTER_WAVES } from "#app/constants/mystery-encounters";
 import i18next from "i18next";
 
 /** the i18n namespace for this encounter */
