+++ src/data/mystery-encounters/encounters/the-winstrate-challenge-encounter.ts
@@ -1,4 +1,4 @@
-import { CLASSIC_MODE_MYSTERY_ENCOUNTER_WAVES } from "#app/constants/constants";
+import { CLASSIC_MODE_MYSTERY_ENCOUNTER_WAVES } from "#app/constants/mystery-encounters";
 import type { PostBattleInitAbAttr } from "#app/data/abilities/ab-attrs/post-battle-init-ab-attr";
 import { applyAbAttrs } from "#app/data/abilities/apply-ab-attrs";
 import type MysteryEncounter from "#app/data/mystery-encounters/mystery-encounter";
